Rename ordered files
