+++ src/sam/02-after/index.js
@@ -1,8 +1,8 @@
 let series = require('run-series')
 let appApex  = require('./00-get-app-apex')
-let deployWS = require('./02-deploy-ws')
+let deployWS = require('./01-deploy-ws')
 let prune = require('../../static')
-let cleanup = require('./03-clean-up-artifacts')
+let cleanup = require('./02-clean-up-artifacts')
 
 module.exports = function after (params, callback) {
   let bucket = undefined // Important, otherwise prune will destroy cfn deployment artifacts instead
